backend/requirements.txt
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.1
Django==2.1.12
django-cors-headers==3.1.1
django-extensions==2.2.5
django-filter==2.0.0
django-formtools==2.1
django-otp==0.7.3
django-phonenumber-field==1.3.0
django-rest-registration==0.5.2
django-two-factor-auth==1.9.1
djangorestframework==3.9.1
drf-yasg==1.10.2
flex==6.14.1
idna==2.7
inflection==0.3.1
ipython==7.9.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonpointer==1.14
jsonschema==2.6.0
Markdown==3.0.1
MarkupSafe==1.0
parso==0.5.1
pexpect==4.7.0
phonenumberslite==8.10.22
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pydotplus==2.0.2
Pygments==2.4.2
pyparsing==2.4.4
pytz==2018.5
PyYAML==5.1.2
qrcode==6.1
requests==2.7.0
rfc3987==1.3.8
ruamel.yaml==0.15.72
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
traitlets==4.3.3
uritemplate==3.0.0
urllib3==1.25.6
validate-email==1.3
wcwidth==0.1.7
